pyproject.toml
[tool.poetry]
name = "pixessa"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0"
djangorestframework = "^3.14.0"
pillow = "^11.1.0"
djangorestframework-simplejwt = "^5.4.0"
dj-rest-auth = {version = "^7.0.1", extras = ["with_social"]}
django-extensions = "^3.2.3"
drf-nested-routers = "^0.94.1"


[tool.poetry.group.dev.dependencies]
bpython = "^0.25"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
